refactor(login): split LoginPostValidate into 3 sub-modules (P12d)
Phase 12d cluster decoupling - wave 3 of 8 files.

What
- Extract from src/Scrapers/Pipeline/Mediator/Login/LoginPostValidate.ts.
- New: PostValidate/{PostValidateDetect,PostValidateGates,PostValidateFlow}.ts.
- LoginPostValidate.ts shrinks to a 50-line thin facade preserving the
  default executeValidateLogin + named detectAsyncLoginErrors exports.
- Add 16 framework-internal helper names to the docs-coverage allowlist
  (sub-modules export to talk to each other; the facade remains the
  sole user-facing surface).

Why
- Continue the post-Phase-12c HIGH-cluster decoupling sweep.
- Same KG-driven pattern that landed Phase 12d wave 1 (LoginScopeIntact,
  commit 232ff71f) and wave 2 (LoginSubmitResolve, commit 6fb15613).

Public-surface contract
- 3 named lib/index.cjs exports preserved.
- 41-cycle baseline preserved (madge --circular).
- All ESLint canaries + guideline clusters preserved.
- Behavior-identical per executeValidateLogin signature audit.

Tracker(s)
Refs Phase 12d (PVTI_lAHOAPtF084BZbzwzgvauEI)
Refs master row Phase 12 (pipeline decoupling)
